Made the page responsive
