+++ frontend/src/App.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import PTMS from './components/PTMS';
-import Login from './components/Login';
+import Login from './components/login/Login';
 
 const App = () => {
   return (
