+++ frontend/src/styles/LoginSelection.css
@@ -1,11 +1,7 @@
 .login-selection {
-  width: 50%;
+  flex: 1;
+  padding: 32px; /* Increased padding for larger devices */
   background-color: white;
-  padding: 48px;
-}
-
-.login-selection-container {
-  height: 100%;
   display: flex;
   flex-direction: column;
   justify-content: center;
@@ -13,37 +9,47 @@
 
 .role-grid {
   display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  gap: 24px;
+  grid-template-columns: repeat(auto-fit, minmax(200px, 1fr)); /* Increased min size for larger screens */
+  gap: 24px; /* Larger gap between cards */
 }
 
 .role-button {
-  padding: 24px;
+  padding: 20px; /* Increased padding for better spacing */
   border-radius: 12px;
-  text-align: left;
-  transition: all 0.3s;
-  text-decoration: none;
-  color: inherit;
+  text-align: center;
+  background-color: #f3f4f6; /* Set background explicitly */
+  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
+  text-decoration: none; /* Remove underline */
+  color: #1d4ed8; /* Ensure the color matches the theme */
+  transition: transform 0.3s;
 }
 
 .role-button:hover {
   transform: translateY(-5px);
-  box-shadow: 0 10px 20px rgba(0, 0, 0, 0.1);
-}
-
-.role-icon {
-  height: 32px;
-  width: 32px;
-  margin-bottom: 12px;
+  box-shadow: 0 6px 20px rgba(0, 0, 0, 0.15);
 }
 
 .role-title {
-  font-size: 1.125rem;
+  font-size: 1.25rem; /* Adjusted for larger screens */
   font-weight: 600;
-  margin-bottom: 4px;
+  margin-bottom: 8px;
 }
 
 .role-description {
-  font-size: 0.875rem;
+  font-size: 1rem;
+  color: #64748b;
 }
 
+@media (max-width: 480px) {
+  .role-grid {
+    gap: 16px;
+  }
+
+  .role-button {
+    padding: 16px;
+  }
+
+  .role-title {
+    font-size: 1rem;
+  }
+}
